+++ src/Recyclarr.TrashLib/Config/Services/ConfigTemplateGuideService.cs
@@ -31,8 +31,10 @@ public class ConfigTemplateGuideService : IConfigTemplateGuideService
         _repo = repo;
     }
 
-    public IReadOnlyCollection<TemplatePath> LoadTemplateData()
+    public async Task<IReadOnlyCollection<TemplatePath>> LoadTemplateData()
     {
+        await _repo.Update();
+
         var templatesPath = _repo.Path.File("templates.json");
         if (!templatesPath.Exists)
         {