+++ src/Recyclarr.TrashLib/Config/Services/IConfigTemplateGuideService.cs
@@ -2,5 +2,5 @@ namespace Recyclarr.TrashLib.Config.Services;
 
 public interface IConfigTemplateGuideService
 {
-    IReadOnlyCollection<TemplatePath> LoadTemplateData();
+    Task<IReadOnlyCollection<TemplatePath>> LoadTemplateData();
 }