+++ src/Recyclarr.TrashLib/TrashLibAutofacModule.cs
@@ -1,5 +1,8 @@
+using System.Reflection;
 using Autofac;
 using Autofac.Extras.Ordering;
+using AutoMapper.Contrib.Autofac.DependencyInjection;
+using AutoMapper.EquivalencyExpression;
 using Recyclarr.Common;
 using Recyclarr.Common.FluentValidation;
 using Recyclarr.TrashLib.ApiServices;
@@ -9,11 +12,14 @@ using Recyclarr.TrashLib.Http;
 using Recyclarr.TrashLib.Repo;
 using Recyclarr.TrashLib.Repo.VersionControl;
 using Recyclarr.TrashLib.Startup;
+using Module = Autofac.Module;
 
 namespace Recyclarr.TrashLib;
 
 public class TrashLibAutofacModule : Module
 {
+    public Assembly? AdditionalMapperProfileAssembly { get; init; }
+
     protected override void Load(ContainerBuilder builder)
     {
         base.Load(builder);
@@ -31,6 +37,14 @@ public class TrashLibAutofacModule : Module
         builder.RegisterModule<ConfigAutofacModule>();
         builder.RegisterType<ServiceRequestBuilder>().As<IServiceRequestBuilder>();
         builder.RegisterType<FlurlClientFactory>().As<IFlurlClientFactory>().SingleInstance();
+
+        var mapperAssemblies = new List<Assembly> {ThisAssembly};
+        if (AdditionalMapperProfileAssembly is not null)
+        {
+            mapperAssemblies.Add(AdditionalMapperProfileAssembly);
+        }
+
+        builder.RegisterAutoMapper(c => c.AddCollectionMappers(), false, mapperAssemblies.ToArray());
     }
 
     private static void CommonRegistrations(ContainerBuilder builder)