+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Clone config template repo when `config create -t` is used.
+
 ## [5.1.0] - 2023-06-26
 
 ### Added